refactored adaptV4Theme
