+++ packages/material-ui/src/styles/index.js
@@ -1,6 +1,6 @@
 export * from './colorManipulator';
 export { default as createMuiTheme } from './createMuiTheme';
-export { default as deprecated_createMuiTheme } from './deprecated_createMuiTheme';
+export { default as adaptV4Theme } from './adaptV4Theme';
 export { default as unstable_createMuiStrictModeTheme } from './createMuiStrictModeTheme';
 export { default as createStyles } from './createStyles';
 export { default as makeStyles } from './makeStyles';
